refactor: [#1258] make things private or pub(crate) when possible.
Limit the exposed funtionality for pacakges. Specially the new
`tracker-core` package which has not been published yet.
