+++ packages/tracker-core/src/authentication/key/repository/in_memory.rs
@@ -9,21 +9,22 @@ pub struct InMemoryKeyRepository {
 
 impl InMemoryKeyRepository {
     /// It adds a new authentication key.
-    pub async fn insert(&self, auth_key: &PeerKey) {
+    pub(crate) async fn insert(&self, auth_key: &PeerKey) {
         self.keys.write().await.insert(auth_key.key.clone(), auth_key.clone());
     }
 
     /// It removes an authentication key.
-    pub async fn remove(&self, key: &Key) {
+    pub(crate) async fn remove(&self, key: &Key) {
         self.keys.write().await.remove(key);
     }
 
-    pub async fn get(&self, key: &Key) -> Option<PeerKey> {
+    pub(crate) async fn get(&self, key: &Key) -> Option<PeerKey> {
         self.keys.read().await.get(key).cloned()
     }
 
     /// It clears all the authentication keys.
-    pub async fn clear(&self) {
+    #[allow(dead_code)]
+    pub(crate) async fn clear(&self) {
         let mut keys = self.keys.write().await;
         keys.clear();
     }
