+++ packages/tracker-core/src/authentication/key/repository/persisted.rs
@@ -21,7 +21,7 @@ impl DatabaseKeyRepository {
     /// # Errors
     ///
     /// Will return a `databases::error::Error` if unable to add the `auth_key` to the database.
-    pub fn add(&self, peer_key: &PeerKey) -> Result<(), databases::error::Error> {
+    pub(crate) fn add(&self, peer_key: &PeerKey) -> Result<(), databases::error::Error> {
         self.database.add_key_to_keys(peer_key)?;
         Ok(())
     }
@@ -31,7 +31,7 @@ impl DatabaseKeyRepository {
     /// # Errors
     ///
     /// Will return a `database::Error` if unable to remove the `key` from the database.
-    pub fn remove(&self, key: &Key) -> Result<(), databases::error::Error> {
+    pub(crate) fn remove(&self, key: &Key) -> Result<(), databases::error::Error> {
         self.database.remove_key_from_keys(key)?;
         Ok(())
     }
@@ -41,7 +41,7 @@ impl DatabaseKeyRepository {
     /// # Errors
     ///
     /// Will return a `database::Error` if unable to load the keys from the database.
-    pub fn load_keys(&self) -> Result<Vec<PeerKey>, databases::error::Error> {
+    pub(crate) fn load_keys(&self) -> Result<Vec<PeerKey>, databases::error::Error> {
         let keys = self.database.load_keys()?;
         Ok(keys)
     }
